+++ vm/natives/Cargo.toml
@@ -11,6 +11,7 @@ errmapgen = { git = "https://github.com/starcoinorg/diem", rev="77c39e00b2030959
 docgen = { git = "https://github.com/starcoinorg/diem", rev="77c39e00b2030959abfc200aeabdfe1096c12a1c" }
 move-command-line-common = { git = "https://github.com/starcoinorg/diem", rev="77c39e00b2030959abfc200aeabdfe1096c12a1c" }
 move-prover = { git = "https://github.com/starcoinorg/diem", rev="77c39e00b2030959abfc200aeabdfe1096c12a1c" }
+move-stdlib = { git = "https://github.com/starcoinorg/diem", rev="77c39e00b2030959abfc200aeabdfe1096c12a1c" }
 move-vm-types = { git = "https://github.com/starcoinorg/diem", rev="77c39e00b2030959abfc200aeabdfe1096c12a1c" }
 move-vm-runtime = { git = "https://github.com/starcoinorg/diem", rev="77c39e00b2030959abfc200aeabdfe1096c12a1c" }
 move-binary-format = { git = "https://github.com/starcoinorg/diem", rev="77c39e00b2030959abfc200aeabdfe1096c12a1c" }
@@ -25,10 +26,11 @@ arrayref = "0.3"
 ripemd160 = "0.9.1"
 num_enum = "0.5.4"
 starcoin-uint = {path = "../../types/uint"}
+starcoin-vm-types=  {path = "../types"}
 
 [dev-dependencies]
 rand = "0.8.4"
 hex = "0.4"
 
 [features]
-testing=[]
+testing=["move-stdlib/testing"]
