+++ vm/stdlib/modules/doc/U256.md
@@ -14,13 +14,13 @@
 -  [Function `from_big_endian`](#0x1_U256_from_big_endian)
 -  [Function `from_little_endian`](#0x1_U256_from_little_endian)
 -  [Function `compare`](#0x1_U256_compare)
--  [Function `add_nocarry`](#0x1_U256_add_nocarry)
--  [Function `sub_noborrow`](#0x1_U256_sub_noborrow)
 -  [Function `add`](#0x1_U256_add)
 -  [Function `sub`](#0x1_U256_sub)
 -  [Function `mul`](#0x1_U256_mul)
 -  [Function `div`](#0x1_U256_div)
 -  [Function `rem`](#0x1_U256_rem)
+-  [Function `add_nocarry`](#0x1_U256_add_nocarry)
+-  [Function `sub_noborrow`](#0x1_U256_sub_noborrow)
 -  [Function `from_bytes`](#0x1_U256_from_bytes)
 -  [Function `native_add`](#0x1_U256_native_add)
 -  [Function `native_sub`](#0x1_U256_native_sub)
@@ -300,77 +300,6 @@ vector should always has two elements.
 
 </details>
 
-<a name="0x1_U256_add_nocarry"></a>
-
-## Function `add_nocarry`
-
-implementation native_add in move.
-
-
-<pre><code><b>fun</b> <a href="U256.md#0x1_U256_add_nocarry">add_nocarry</a>(a: &<b>mut</b> <a href="U256.md#0x1_U256_U256">U256::U256</a>, b: &<a href="U256.md#0x1_U256_U256">U256::U256</a>)
-</code></pre>
-
-
-
-<details>
-<summary>Implementation</summary>
-
-
-<pre><code><b>fun</b> <a href="U256.md#0x1_U256_add_nocarry">add_nocarry</a>(a: &<b>mut</b> <a href="U256.md#0x1_U256">U256</a>, b: &<a href="U256.md#0x1_U256">U256</a>) {
-    <b>let</b> carry = 0;
-    <b>let</b> idx = 0;
-    <b>let</b> len = (<a href="U256.md#0x1_U256_WORD">WORD</a> <b>as</b> u64);
-    <b>while</b> (idx &lt; len) {
-        <b>let</b> a_bit = <a href="Vector.md#0x1_Vector_borrow_mut">Vector::borrow_mut</a>(&<b>mut</b> a.bits, idx);
-        <b>let</b> b_bit = <a href="Vector.md#0x1_Vector_borrow">Vector::borrow</a>(&b.bits, idx);
-        *a_bit = <a href="U256.md#0x1_Arith_adc">0x1::Arith::adc</a>(*a_bit, *b_bit, &<b>mut</b> carry);
-    };
-
-    // check overflow
-    <b>assert</b>(carry == 0, 100);
-}
-</code></pre>
-
-
-
-</details>
-
-<a name="0x1_U256_sub_noborrow"></a>
-
-## Function `sub_noborrow`
-
-implementation native_sub in move.
-
-
-<pre><code><b>fun</b> <a href="U256.md#0x1_U256_sub_noborrow">sub_noborrow</a>(a: &<b>mut</b> <a href="U256.md#0x1_U256_U256">U256::U256</a>, b: &<a href="U256.md#0x1_U256_U256">U256::U256</a>)
-</code></pre>
-
-
-
-<details>
-<summary>Implementation</summary>
-
-
-<pre><code><b>fun</b> <a href="U256.md#0x1_U256_sub_noborrow">sub_noborrow</a>(a: &<b>mut</b> <a href="U256.md#0x1_U256">U256</a>, b: &<a href="U256.md#0x1_U256">U256</a>) {
-    <b>let</b> borrow = 0;
-    <b>let</b> idx = 0;
-    <b>let</b> len =(<a href="U256.md#0x1_U256_WORD">WORD</a> <b>as</b> u64);
-    <b>while</b> (idx &lt; len) {
-        <b>let</b> a_bit = <a href="Vector.md#0x1_Vector_borrow_mut">Vector::borrow_mut</a>(&<b>mut</b> a.bits, idx);
-        <b>let</b> b_bit = <a href="Vector.md#0x1_Vector_borrow">Vector::borrow</a>(&b.bits, idx);
-        *a_bit = <a href="U256.md#0x1_Arith_sbb">0x1::Arith::sbb</a>(*a_bit, *b_bit, &<b>mut</b> borrow);
-    };
-
-    // check overflow
-    <b>assert</b>(borrow == 0, 100);
-
-}
-</code></pre>
-
-
-
-</details>
-
 <a name="0x1_U256_add"></a>
 
 ## Function `add`
@@ -387,7 +316,7 @@ implementation native_sub in move.
 
 
 <pre><code><b>public</b> <b>fun</b> <a href="U256.md#0x1_U256_add">add</a>(a: <a href="U256.md#0x1_U256">U256</a>, b: <a href="U256.md#0x1_U256">U256</a>): <a href="U256.md#0x1_U256">U256</a> {
-    <a href="U256.md#0x1_U256_native_add">native_add</a>(&<b>mut</b> a, &b);
+    <a href="U256.md#0x1_U256_add_nocarry">add_nocarry</a>(&<b>mut</b> a, &b);
     a
 }
 </code></pre>
@@ -412,7 +341,7 @@ implementation native_sub in move.
 
 
 <pre><code><b>public</b> <b>fun</b> <a href="U256.md#0x1_U256_sub">sub</a>(a: <a href="U256.md#0x1_U256">U256</a>, b: <a href="U256.md#0x1_U256">U256</a>): <a href="U256.md#0x1_U256">U256</a> {
-    <a href="U256.md#0x1_U256_native_sub">native_sub</a>(&<b>mut</b> a, &b);
+    <a href="U256.md#0x1_U256_sub_noborrow">sub_noborrow</a>(&<b>mut</b> a, &b);
     a
 }
 </code></pre>
@@ -489,6 +418,79 @@ implementation native_sub in move.
 <pre><code><b>public</b> <b>fun</b> <a href="U256.md#0x1_U256_rem">rem</a>(a: <a href="U256.md#0x1_U256">U256</a>, b: <a href="U256.md#0x1_U256">U256</a>): <a href="U256.md#0x1_U256">U256</a> {
     <a href="U256.md#0x1_U256_native_rem">native_rem</a>(&<b>mut</b> a, &b);
     a
+}
+</code></pre>
+
+
+
+</details>
+
+<a name="0x1_U256_add_nocarry"></a>
+
+## Function `add_nocarry`
+
+move implementation of native_add.
+
+
+<pre><code><b>fun</b> <a href="U256.md#0x1_U256_add_nocarry">add_nocarry</a>(a: &<b>mut</b> <a href="U256.md#0x1_U256_U256">U256::U256</a>, b: &<a href="U256.md#0x1_U256_U256">U256::U256</a>)
+</code></pre>
+
+
+
+<details>
+<summary>Implementation</summary>
+
+
+<pre><code><b>fun</b> <a href="U256.md#0x1_U256_add_nocarry">add_nocarry</a>(a: &<b>mut</b> <a href="U256.md#0x1_U256">U256</a>, b: &<a href="U256.md#0x1_U256">U256</a>) {
+    <b>let</b> carry = 0;
+    <b>let</b> idx = 0;
+    <b>let</b> len = (<a href="U256.md#0x1_U256_WORD">WORD</a> <b>as</b> u64);
+    <b>while</b> (idx &lt; len) {
+        <b>let</b> a_bit = <a href="Vector.md#0x1_Vector_borrow_mut">Vector::borrow_mut</a>(&<b>mut</b> a.bits, idx);
+        <b>let</b> b_bit = <a href="Vector.md#0x1_Vector_borrow">Vector::borrow</a>(&b.bits, idx);
+        *a_bit = <a href="U256.md#0x1_Arith_adc">0x1::Arith::adc</a>(*a_bit, *b_bit, &<b>mut</b> carry);
+        idx = idx + 1;
+    };
+
+    // check overflow
+    <b>assert</b>(carry == 0, 100);
+}
+</code></pre>
+
+
+
+</details>
+
+<a name="0x1_U256_sub_noborrow"></a>
+
+## Function `sub_noborrow`
+
+move implementation of native_sub.
+
+
+<pre><code><b>fun</b> <a href="U256.md#0x1_U256_sub_noborrow">sub_noborrow</a>(a: &<b>mut</b> <a href="U256.md#0x1_U256_U256">U256::U256</a>, b: &<a href="U256.md#0x1_U256_U256">U256::U256</a>)
+</code></pre>
+
+
+
+<details>
+<summary>Implementation</summary>
+
+
+<pre><code><b>fun</b> <a href="U256.md#0x1_U256_sub_noborrow">sub_noborrow</a>(a: &<b>mut</b> <a href="U256.md#0x1_U256">U256</a>, b: &<a href="U256.md#0x1_U256">U256</a>) {
+    <b>let</b> borrow = 0;
+    <b>let</b> idx = 0;
+    <b>let</b> len =(<a href="U256.md#0x1_U256_WORD">WORD</a> <b>as</b> u64);
+    <b>while</b> (idx &lt; len) {
+        <b>let</b> a_bit = <a href="Vector.md#0x1_Vector_borrow_mut">Vector::borrow_mut</a>(&<b>mut</b> a.bits, idx);
+        <b>let</b> b_bit = <a href="Vector.md#0x1_Vector_borrow">Vector::borrow</a>(&b.bits, idx);
+        *a_bit = <a href="U256.md#0x1_Arith_sbb">0x1::Arith::sbb</a>(*a_bit, *b_bit, &<b>mut</b> borrow);
+        idx = idx + 1;
+    };
+
+    // check overflow
+    <b>assert</b>(borrow == 0, 100);
+
 }
 </code></pre>
