+++ vm/stdlib/tests/stdlib_unit_test.rs
@@ -1,6 +1,13 @@
 use move_unit_test::UnitTestingConfig;
+use starcoin_natives::starcoin_natives;
 move_unit_test::register_move_unit_tests!(
-    UnitTestingConfig::default_with_bound(Some(100_000_000)),
-    "../modules",
-    r".*\.move$"
+    {
+        let mut c = UnitTestingConfig::default_with_bound(Some(100_000_000));
+        c.verbose = true;
+        c
+    },
+    "./modules",
+    r".*\.move$",
+    "./src",
+    Some(starcoin_natives())
 );
